ISSUE #4133 - Added Upon log in, you should be redirected to the page you wish to go test
